requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
dymo-bluetooth>=0.1.3
pillow>=10.0
python-barcode>=0.15
qrcode[pil]>=7.4
aiofiles>=23.0
python-multipart>=0.0.9
